requirements.txt
py-cord
python-dotenv
requests
feedparser
instagrapi
Pillow